Move 'versionAs...' tasks to root
