+++ .github/workflows/node-zxc-build-release-artifact.yaml
@@ -196,13 +196,13 @@ jobs:
         uses: gradle/gradle-build-action@3fbe033aaae657f011f88f29be9e65ed26bd29ef # pin@v2
         if: ${{ inputs.version-policy == 'specified' && !cancelled() && !failure() }}
         with:
-          arguments: -p hedera-node versionAsSpecified -PnewVersion=${{ inputs.new-version }} --scan
+          arguments: versionAsSpecified -PnewVersion=${{ inputs.new-version }} --scan
 
       - name: Gradle Update Version (Branch Commit)
         uses: gradle/gradle-build-action@3fbe033aaae657f011f88f29be9e65ed26bd29ef # pin@v2
         if: ${{ inputs.version-policy != 'specified' && !cancelled() && !failure() }}
         with:
-          arguments: -p hedera-node versionAsPrefixedCommit -PcommitPrefix=${{ steps.parameters.outputs.commit-prefix }} --scan
+          arguments: versionAsPrefixedCommit -PcommitPrefix=${{ steps.parameters.outputs.commit-prefix }} --scan
 
       - name: Gradle Version Summary
         uses: gradle/gradle-build-action@3fbe033aaae657f011f88f29be9e65ed26bd29ef # pin@v2
