+++ .github/workflows/platform-zxc-release-internal.yaml
@@ -208,21 +208,21 @@ jobs:
         if: ${{ (inputs.version-policy == 'uniform-explicit' || inputs.version-policy == 'early-access-explicit') && !cancelled() && !failure() }}
         with:
           gradle-version: ${{ inputs.gradle-version }}
-          arguments: -p platform-sdk versionAsSpecified --scan -PnewVersion=${{ steps.verify-explicit-version.outputs.number }}
+          arguments: versionAsSpecified --scan -PnewVersion=${{ steps.verify-explicit-version.outputs.number }}
 
       - name: Update Version Numbers (Commit Locked)
         if: ${{ inputs.version-policy == 'commit-locked' && !cancelled() && !failure() }}
         uses: gradle/gradle-build-action@v2
         with:
           gradle-version: ${{ inputs.gradle-version }}
-          arguments: -p platform-sdk versionAsPrefixedCommit --scan -PcommitPrefix=${{ inputs.prerelease-prefix }}
+          arguments: versionAsPrefixedCommit --scan -PcommitPrefix=${{ inputs.prerelease-prefix }}
 
       - name: Update Version Numbers (Snapshot Locked)
         if: ${{ inputs.version-policy == 'snapshot-locked' && !cancelled() && !failure() }}
         uses: gradle/gradle-build-action@v2
         with:
           gradle-version: ${{ inputs.gradle-version }}
-          arguments: -p platform-sdk versionAsSnapshot --scan
+          arguments: versionAsSnapshot --scan
 
       - name: Version Report
         uses: gradle/gradle-build-action@v2
