+++ build-logic/project-plugins/src/main/kotlin/com.hedera.hashgraph.java.gradle.kts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+import Utils.Companion.versionTxt
 import com.autonomousapps.AbstractExtension
 import com.autonomousapps.DependencyAnalysisSubExtension
 
@@ -171,8 +172,3 @@ val dependencyAnalysis = extensions.findByType<AbstractExtension>()
 if (dependencyAnalysis is DependencyAnalysisSubExtension) {
     dependencyAnalysis.issues { onAny { exclude(project.path) } }
 }
-
-// Find the version.txt in the root of the repository, independent of
-// which build is started from where.
-fun Directory.versionTxt(): RegularFile =
-    file("version.txt").let { if (it.asFile.exists()) it else this.dir("..").versionTxt() }
