+++ build-logic/project-plugins/src/main/kotlin/com.hedera.hashgraph.umbrella.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016-2023 Hedera Hashgraph, LLC
+ * Copyright (C) 2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,15 @@
  * limitations under the License.
  */
 
+import Utils.Companion.versionTxt
+import net.swiftzer.semver.SemVer
+
 plugins {
     id("lifecycle-base")
     id("com.hedera.hashgraph.repositories")
     id("com.hedera.hashgraph.spotless-conventions")
     id("com.hedera.hashgraph.spotless-kotlin-conventions")
+    id("lazy.zoo.gradle.git-data-plugin")
 }
 
 spotless { kotlinGradle { target("build-logic/**/*.gradle.kts") } }
@@ -57,5 +61,71 @@ fun configureLifecycleTask(taskName: String) {
                 .filter { it.name != "build-logic" }
                 .map { build -> build.task(":${taskName}") }
         )
+    }
+}
+
+tasks.register("githubVersionSummary") {
+    group = "github"
+    doLast {
+        val ghStepSummaryPath: String =
+            providers.environmentVariable("GITHUB_STEP_SUMMARY").orNull
+                ?: throw IllegalArgumentException(
+                    "This task may only be run in a Github Actions CI environment!" +
+                        "Unable to locate the GITHUB_STEP_SUMMARY environment variable."
+                )
+
+        Utils.generateProjectVersionReport(
+            rootProject,
+            File(ghStepSummaryPath).outputStream().buffered()
+        )
+    }
+}
+
+tasks.register("showVersion") {
+    group = "versioning"
+    doLast { println(project.version) }
+}
+
+tasks.register("versionAsPrefixedCommit") {
+    group = "versioning"
+    doLast {
+        gitData.lastCommitHash?.let {
+            val prefix = providers.gradleProperty("commitPrefix").getOrElse("adhoc")
+            val newPrerel = prefix + ".x" + it.take(8)
+            val currVer =
+                SemVer.parse(layout.projectDirectory.versionTxt().asFile.readText().trim())
+            try {
+                val newVer = SemVer(currVer.major, currVer.minor, currVer.patch, newPrerel)
+                Utils.updateVersion(rootProject, newVer)
+            } catch (e: java.lang.IllegalArgumentException) {
+                throw IllegalArgumentException(String.format("%s: %s", e.message, newPrerel), e)
+            }
+        }
+    }
+}
+
+tasks.register("versionAsSnapshot") {
+    group = "versioning"
+    doLast {
+        val currVer = SemVer.parse(layout.projectDirectory.versionTxt().asFile.readText().trim())
+        val newVer = SemVer(currVer.major, currVer.minor, currVer.patch, "SNAPSHOT")
+
+        Utils.updateVersion(rootProject, newVer)
+    }
+}
+
+tasks.register("versionAsSpecified") {
+    group = "versioning"
+    doLast {
+        val verStr = providers.gradleProperty("newVersion")
+
+        if (!verStr.isPresent) {
+            throw IllegalArgumentException(
+                "No newVersion property provided! Please add the parameter -PnewVersion=<version> when running this task."
+            )
+        }
+
+        val newVer = SemVer.parse(verStr.get())
+        Utils.updateVersion(rootProject, newVer)
     }
 }
